Max health stat

+++ src/main/java/io/github/stuff_stuffs/turnbasedcombat/common/battle/entity/EntityStateView.java
@@ -5,7 +5,7 @@ import io.github.stuff_stuffs.turnbasedcombat.common.battle.Team;
 import java.util.UUID;
 
 public interface EntityStateView {
-    int getHealth();
+    double getHealth();
 
     int getLevel();
